% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_PIs.R
\name{compute_PIs}
\alias{compute_PIs}
\title{Compute the prediction intervals for responses of new test points from a
fitted generalized partial linear spatially varying coefficient model.}
\usage{
compute_PIs(
  Y_train,
  X_train,
  ind_l,
  ind_nl,
  U_train,
  X_pred,
  U_pred,
  V,
  Tr,
  d = 2,
  r = 1,
  lambda = 10^seq(-6, 6, by = 0.5),
  family,
  off = 0,
  r_theta = c(2, 8),
  eps = 0.01,
  method = "CV+",
  cp = 0.95,
  nfold = 10
)
}
\arguments{
\item{Y_train}{The response variable,a \code{n} by one matrix where \code{n}
is the number of observations in the training data set .}

\item{X_train}{The design matrix of \code{n} by \code{np} where \code{np} is
the number of covariates (including intercept if intercept exists). Each row is a vector of the covariates for an
observation in the training data set.}

\item{ind_l}{The vector of the indexes which indicate the columns of linear
covariates in \code{X_train}.}

\item{ind_nl}{The vector of the indexes which indicate the columns of
nonlinear covariates in \code{X_train}.}

\item{U_train}{A \code{n} by two matrix where each row is the coordinates of
an observation in the training data set.}

\item{X_pred}{The design matrix for prediction (including intercept if intercept exists).}

\item{U_pred}{The matrix of coordinates for prediction.}

\item{V}{A \code{N} by two matrix of vertices of a triangulation, where
\code{N} is the number of vertices and each row is the coordinates for a
vertex.}

\item{Tr}{A \code{n_Tr} by three triangulation matrix, where \code{n_Tr} is
the number of triangles in the triangulation and each row is the indices of
vertices in \code{V}.}

\item{d}{The degree of piecewise polynomials -- default is 2.}

\item{r}{The smoothness parameter and \code{r} \eqn{<} \code{d} -- default is
1.}

\item{lambda}{The vector of the candidates of penalty parameter -- default is
grid points of 10 to the power of a sequence from -6 to 6 by 0.5.}

\item{family}{The family object which specifies the distribution and link to
use (see \code{\link{glm}} and \code{\link{family}}).}

\item{off}{The offset -- default is 0.}

\item{r_theta}{The vector of the upper and lower bound of an interval to
search for an additional parameter \code{theta} for negative binomial
scenario -- default is c(2,8).}

\item{eps}{The error tolerance for the Pearson estimate of the scale
parameter, which is as close  to 1, when estimating an additional parameter
\code{theta} for negative binomial scenario -- default is 0.01.}

\item{method}{The prediction method used in the computation, options are
"CV+", "Jackknife" and "Jackknife+" -- default is "CV+".}

\item{cp}{The desired coverage level for the prediction intervals -- default
is 0.95.}

\item{nfold}{The number of folds for CV+ method -- default is 10.}
}
\value{
A data frame of computed prediction intervals for responses of the
  predicted points.
}
\description{
\code{compute_PIs} compute the prediction intervals for responses of new test
points from a fitted \code{gplsvcm} object based on a selected prediction
method among Jackknife, Jackknife+ and K-fold cross validation (CV+), and
return the prediction intervals for responses of predicted points.
}
\details{
The construction of the polynomial spline functions is via
  \code{\link[BPST]{basis}}.
}
\examples{
# See an example of gplsvcm_fitwMIDF.

}
\references{
Barber et al.(2021) Predictive inference with the jackknife+.
  Ann.Statist.49(1):486-507
  \url{https://projecteuclid.org/journals/annals-of-statistics/volume-49/issue-1/Predictive-inference-with-the-jackknife/10.1214/20-AOS1965.full}
}
