% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_fit.R
\name{CV_fit}
\alias{CV_fit}
\title{Fit the model with K fold cross validation and compute the CV residuals}
\usage{
CV_fit(
  Y_train,
  X_l_train,
  X_nl_train,
  ind_l,
  ind_nl,
  U_train,
  X_l_pred,
  X_nl_pred,
  U_pred,
  B_pred,
  V,
  Tr,
  d,
  r,
  lambda,
  family,
  off,
  r_theta,
  eps,
  nfold
)
}
\description{
This is an internal function of package \code{GPLSVCM}.
}
