% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_gplsvcm.R
\name{cv_gplsvcm}
\alias{cv_gplsvcm}
\title{Calculate the K-fold Cross Validation Mean Square Prediction Error from a
fitted generalized partial linear spatially varying coefficient model.}
\usage{
cv_gplsvcm(
  Y,
  X,
  ind_l,
  ind_nl,
  U,
  V,
  Tr,
  d = 2,
  r = 1,
  lambda = 10^seq(-6, 6, by = 0.5),
  family,
  off = 0,
  r_theta = c(2, 8),
  eps = 0.01,
  nfold = 10
)
}
\arguments{
\item{Y}{The response variable,a \code{n} by one matrix where \code{n} is the
number of observations.}

\item{X}{The design matrix of \code{n} by \code{np} where \code{np} is the
number of covariates (including intercept if intercept exists). Each row is a vector of the covariates for an
observation.}

\item{ind_l}{The vector of the indexes which indicate the columns of linear
covariates in \code{X}.}

\item{ind_nl}{The vector of the indexes which indicate the columns of
nonlinear covariates in \code{X}.}

\item{U}{A \code{n} by two matrix where each row is the coordinates of an
observation.}

\item{V}{A \code{N} by two matrix of vertices of a triangulation, where
\code{N} is the number of vertices and each row is the coordinates for a
vertex.}

\item{Tr}{A \code{n_Tr} by three triangulation matrix, where \code{n_Tr} is
the number of triangles in the triangulation and each row is the indices of
vertices in \code{V}.}

\item{d}{The degree of piecewise polynomials -- default is 2.}

\item{r}{The smoothness parameter and \code{r} \eqn{<} \code{d} -- default is
1.}

\item{lambda}{The vector of the candidates of penalty parameter -- default is
grid points of 10 to the power of a sequence from -6 to 6 by 0.5.}

\item{family}{The family object which specifies the distribution and link to
use (see \code{\link{glm}} and \code{\link{family}}).}

\item{off}{The offset -- default is 0.}

\item{r_theta}{The vector of the upper and lower bound of an interval to
search for an additional parameter \code{theta} for negative binomial
scenario -- default is c(2,8).}

\item{eps}{The error tolerance for the Pearson estimate of the scale
parameter, which is as close  to 1, when estimating an additional parameter
\code{theta} for negative binomial scenario -- default is 0.01.}

\item{nfold}{The number of folds for cross validation -- default is 10.}
}
\value{
The mean square prediction error (MSPE).
}
\description{
\code{cv_gplsvcm} implements K-fold cross-validation from a fitted
\code{gplsvcm} object, and returns the mean squared prediction error (MSPE).
}
\details{
The construction of the polynomial spline functions is via
  \code{\link[BPST]{basis}}.
}
\examples{
# See an example of gplsvcm_fitwMIDF.

}
