% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.dc.R
\name{dat.dc}
\alias{dat.dc}
\title{gplSVCM: generalized partially linear model with spatially varying coefficients}
\usage{
dat.dc(Vs, Trs, Y, X, Z, S)
}
\description{
This function identifies points falling within a triangulation.
}
