% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.generator.R
\name{dat.generator}
\alias{dat.generator}
\title{gplSVCM: generalized partially linear model with spatially varying coefficients}
\usage{
dat.generator(n, V, Tr, family.name = "gau", seed = 2022)
}
\arguments{
\item{n}{Number of observation points to be generated.}

\item{V}{The nV by two matrix of vertices of a triangulation, where nV is the
number of vertices. Each row is the coordinates for a vertex.}

\item{Tr}{The triangulation matrix of dimention nT by three, where nT is the
number of triangles in the triangulation. Each row is the indices of vertices in V.}

\item{family.name}{A family object specifying the distribution and link to use.}

\item{seed}{An integer of random seed.}

\item{sig}{Standard deviation of the random error in gaussian distribution.}
}
\value{
\item{dat}{A list of model components in a simlulated dataset.}
}
\description{
This function generates generate a dataset based on simulation studies.
}
