% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datagenerator.R
\name{Datagenerator}
\alias{Datagenerator}
\title{Generating populations for simulation.}
\usage{
Datagenerator(family, ngrid)
}
\arguments{
\item{family}{The family object, specifying the distribution and link to use.
Choose "gaussian" for Gaussian distribution, "poisson" for poisson
distribution, and "nb_bps" for negative binomial distribution.}

\item{ngrid}{The distance between grid points -- default is set to 0.02.}
}
\value{
A data matrix with a response ('y'), true coefficient functions ('m1'
  and 'm2'), nonlinear covariates ('x1' and 'x2'), linear covariates ('x3'
  and 'x4') and locations ('u' and 'v').
}
\description{
\code{Datagenerator} is used to generate samples on horseshoe domain for
Scenario 1 (Gaussian), Scenario 2 (Poisson) and Scenario 3 (negative
binomial).
}
\details{
This function used the package \code{mgcv}, see
  \code{\link[mgcv]{fs.boundary}} and \code{\link[mgcv]{fs.test}}
}
\examples{
family=nb_bps()
ngrid = 0.02
pop = Datagenerator(family, ngrid)

}
