% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_est.R
\name{dev_est}
\alias{dev_est}
\title{Calculate the generalized deviance of a fitted model}
\usage{
dev_est(X_l, X_nl, mfit, family)
}
\arguments{
\item{X_l}{The matrix of linear covariates for observations.}

\item{X_nl}{The matrix of nonlinear covariates for observations.}

\item{mfit}{A list of information returned by function \code{gplsvcm_est}}

\item{family}{The family object which specifies the distribution and link to
use.}
}
\description{
\code{dev_est} is an internal function of \code{gplsvcm_fitwMIDF}.
}
