% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsvcm_est.R
\name{gplsvcm_est}
\alias{gplsvcm_est}
\title{Estimation for GPLSCVMs}
\usage{
gplsvcm_est(Y, X_l, X_nl, U, V, Tr, d, r, lambda, family, off, r_theta, eps)
}
\arguments{
\item{Y}{The response variable,a \code{n} by one matrix where \code{n} is the
number of observations.}

\item{X_l}{The matrix of linear covariates for observations, of dimension
\code{n} by \code{np_l} where \code{n} is number of observations and
\code{np_l} is the number of linear covarites.}

\item{X_nl}{The matrix of nonlinear covariates for observations, of dimension
\code{n} by \code{np_nl} where \code{n} is number of observations and
\code{np_nl} is the number of nonlinear covarites.}

\item{U}{A \code{n} by two matrix where each row is the coordinates of an
observation.}

\item{V}{A \code{N} by two matrix of vertices of a triangulation, where
\code{N} is the number of vertices and each row is the coordinates for a
vertex.}

\item{Tr}{A \code{n_Tr} by three triangulation matrix, where \code{n_Tr} is
the number of triangles in the triangulation and each row is the indices of
vertices in \code{V}.}

\item{d}{The degree of piecewise polynomials.}

\item{r}{The smoothness parameter and \code{r} \eqn{<} \code{d}.}

\item{lambda}{The vector of the candidates of penalty parameter.}

\item{family}{The family object which specifies the distribution and link to
use (see \code{\link{glm}} and \code{\link{family}}).}

\item{off}{The offset.}

\item{r_theta}{The vector of the upper and lower bound of an interval to
search for an additional parameter \code{theta} for negative binomial
scenario.}

\item{eps}{The error tolerance for the Pearson estimate of the scale
parameter, which is as close as possible to 1, when estimating an
additional parameter \code{theta} for negative binomial scenario.}
}
\description{
This is an internal function of \code{GPLSVCM} which is used in function
\code{gplsvcm_fit} and \code{gplsvcm_fitwMIDF}.
}
\details{
The construction of the polynomial spline functions is via
  \code{\link[BPST]{basis}}.
}
