% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsvcm_fitwMIDF.R
\name{gplsvcm_fitwMIDF}
\alias{gplsvcm_fitwMIDF}
\title{Fitting the generalized partial linear spatially varying coefficient model
with Model Selection}
\usage{
gplsvcm_fitwMIDF(
  Y,
  X,
  U,
  V,
  Tr,
  d = 2,
  r = 1,
  lambda = 10^seq(-6, 6, by = 0.5),
  family,
  k_n = NULL,
  method = "BIC",
  off = 0,
  r_theta = c(2, 8),
  eps = 0.01
)
}
\arguments{
\item{Y}{The response variable,a \code{n} by one matrix where \code{n} is the
number of observations.}

\item{X}{The design matrix of \code{n} by \code{np} where \code{np} is the
number of covariates (including intercept if intercept exists). Each row is a vector of the covariates for an
observation.}

\item{U}{A \code{n} by two matrix where each row is the coordinates of an
observation.}

\item{V}{A \code{N} by two matrix of vertices of a triangulation, where
\code{N} is the number of vertices and each row is the coordinates for a
vertex.}

\item{Tr}{A \code{n_Tr} by three triangulation matrix, where \code{n_Tr} is
the number of triangles in the triangulation and each row is the indices of
vertices in \code{V}.}

\item{d}{The degree of piecewise polynomials -- default is 2.}

\item{r}{The smoothness parameter and \code{r} \eqn{<} \code{d} -- default is
1.}

\item{lambda}{The vector of the candidates of penalty parameter -- default is
grid points of 10 to the power of a sequence from -6 to 6 by 0.5.}

\item{family}{The family object which specifies the distribution and link to
use (see \code{\link{glm}} and \code{\link{family}}).}

\item{k_n}{The penalty parameter used in the model selection criteria. It
need to be supplied only when the argument \code{method} is set to
\code{NULL} --default is NULL.}

\item{method}{The type of model selection criteria, options are "AIC","BIC"
and NULL which correspond to \code{k_n=2}, \code{k_n=log(n)} and
\code{k_n=k_n} respectively -- default is "BIC".}

\item{off}{The offset -- default is 0.}

\item{r_theta}{The vector of the upper and lower bound of an interval to
search for an additional parameter \code{theta} for negative binomial
scenario -- default is c(2,8).}

\item{eps}{The error tolerance for the Pearson estimate of the scale
parameter, which is as close as possible to 1, when estimating an
additional parameter \code{theta} for negative binomial scenario -- default
is 0.01.}
}
\value{
The function returns a list of fitted object information from S3
  class "gplsvcm", see the items of the list from \code{\link{gplsvcm_fit}}.
}
\description{
\code{gplsvcm_fitwMIDF} perform a model selection procedure to identify the
linear and nonlinear components first and then fit the corresponding
generalized partial linear spatially varying coefficient model.
}
\details{
The \code{gplsvcm_fitwMIDF} function is used to fit a generalized
  partial linear spatially varying coefficient model when the linear and
  nonlinear parts of the design matrix \code{X} are not known before
  analysis. The construction of the polynomial spline functions is via
  \code{\link[BPST]{basis}}. It first perform a model selection based on
  Generalized Information Criterion (GIC) and output the selected model by
  specifying the parameters \code{ind_l} and \code{ind_nl} of the function
  \code{gplsvcm_fit}. Then the selected model is fitted by the function
  \code{gplsvcm_fit}.
}
\examples{
# Population:
family=poisson()
ngrid = 0.02

# Data generation:
pop = Datagenerator(family, ngrid)
N=nrow(pop)

# Triangulations and setup:
Tr = Tr0; V = V0; n = 1000; d = 2; r = 1;

# set up for smoothing parameters in the penalty term:
lambda_start=0.0001; lambda_end=10; nlambda=10;
lambda=exp(seq(log(lambda_start),log(lambda_end),length.out=nlambda))

# Generate Sample:
ind_s=sample(N,n,replace=FALSE)
data=as.matrix(pop[ind_s,])
Y=data[,1]; X=data[,c(6:9)]; U=data[,c(10:11)];

# True coefficents
alpha=data[,c(2:3)]; beta=data[,c(4:5)];

# Fit the model with model selection based on AIC:
mfit1 = gplsvcm_fitwMIDF(Y, X, U, V, Tr, d , r , lambda,family,k_n=NULL,
method="AIC",off = 0,r_theta = c(2, 8), eps= 0.01)

# Fit the model with model selection based on BIC:
mfit2 = gplsvcm_fitwMIDF(Y, X, U, V, Tr, d , r , lambda,family,k_n=NULL,
method="BIC",off = 0,r_theta = c(2, 8), eps= 0.01)

# prediction intervals:
ind_l=mfit2$ind_l; ind_nl=mfit2$ind_nl;
set.seed(123)
PIs=compute_PIs(Y,X,ind_l,ind_nl,U,X,U,V,Tr,d,r,lambda,family,off = 0,
r_theta = c(2, 8), eps= 0.01,method="CV+", cp=0.95, nfold = 10)

# prediction:
Y_hat = gplsvcm_predict(mfit2, X, U)

# k-fold cross-validation:
set.seed(123)
MSPE = cv_gplsvcm(Y,X,ind_l,ind_nl,U,V,Tr,d,r,lambda,family,off = 0,r_theta =
c(2, 8), eps= 0.01,nfold=10)

# plot the estimated coefficients
gplsvcm_plot(mfit2,gridnumber=100,display=c(1,1),xlab=c("u1","u1"),
ylab=c("u2","u2"),main=c(expression(paste("The Estimated Surface for","
",hat(alpha)[1])),expression(paste("The Estimated Surface for","
",hat(alpha)[2]))))

}
\references{
Zhang et al.(2010) Regularization Parameter Selections via
  Generalized Information
  Criterion.\url{https://www.tandfonline.com/doi/abs/10.1198/jasa.2009.tm08013}
}
