% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsvcm_plot.R
\name{gplsvcm_plot}
\alias{gplsvcm_plot}
\title{Produces coefficient function plots for a fitted generalized partial linear
spatially varying coefficient model.}
\usage{
gplsvcm_plot(
  mfit,
  gridnumber = 100,
  display = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  ...
)
}
\arguments{
\item{mfit}{A fitted \code{gplsvcm} object returned from function
\code{gplsvcm_fit} or \code{gplsvcm_fitwMIDF}.}

\item{gridnumber}{The number of grid points on one range for plots -- default
is 100.}

\item{display}{If supplied then it is the vector for specifying how to
display the estimated surfaces for the coefficient functions, used in
\code{par(mfrow=)}.}

\item{xlab}{If supplied then is the vector of characters where each element
is the x label for the estimated surface of one coefficient function.}

\item{ylab}{If supplied then is the vector of characters where each element
is the y label for the estimated surface of one coefficient function.}

\item{main}{If supplied then is the vector of characters where each element
is the title for the estimated surface of one coefficient function.}

\item{...}{other graphics parameters to pass on to plotting commands. See
details in \code{\link[plot3D]{image2D}}.}
}
\value{
None
}
\description{
\code{gplsvcm_plot} produces the plots of the estimated coefficient functions
from a fitted \code{gplsvcm} object.
}
\details{
This function used package \code{Triangulation} and \code{plot3D},
  see \code{\link[Triangulation]{TriPlot}} and \code{\link[plot3D]{image2D}}.
}
\examples{
# See an example of gplsvcm_fitwMIDF.

}
