% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplsvcm_predict.R
\name{gplsvcm_predict}
\alias{gplsvcm_predict}
\title{Predictions for responses of new test points from a fitted generalized
partial linear spatially varying coefficient model.}
\usage{
gplsvcm_predict(mfit, Xpred, Upred)
}
\arguments{
\item{mfit}{A fitted \code{gplsvcm} object returned from function
\code{gplsvcm_fit} or \code{gplsvcm_fitwMIDF}.}

\item{Xpred}{The design matrix for prediction (including intercept if intercept exists).}

\item{Upred}{The matrix of cooridinates for prediction.}
}
\value{
A vector of predicted response.
}
\description{
\code{gplsvcm_predict} is used to make predictions for the responses of
predicted points from a fitted \code{gplsvcm} object.
}
\details{
The construction of the polynomial spline functions is via
  \code{\link[BPST]{basis}}
}
\examples{
# See an example of gplsvcm_fitwMIDF.

}
