% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_fit.R
\name{loo_fit}
\alias{loo_fit}
\title{Fit the model with the i-th training data point removed and compute the leave
one out residuals}
\usage{
loo_fit(
  Y_train,
  X_l_train,
  X_nl_train,
  ind_l,
  ind_nl,
  U_train,
  X_l_pred,
  X_nl_pred,
  U_pred,
  B_pred,
  V,
  Tr,
  d,
  r,
  lambda,
  family,
  off,
  r_theta,
  eps
)
}
\description{
This is an internal function of package \code{GPLSVCM}.
}
