% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfit.gsvcm.R
\name{mfit.gsvcm}
\alias{mfit.gsvcm}
\title{gplSVCM: Generalized Partially Linear Spatially Varying Coefficients Models}
\usage{
mfit.gsvcm(V, Tr, d, r, Y, X, S, family)
}
\arguments{
\item{V}{The nV by two matrix of vertices of a triangulation, where nV is the
number of vertices. Each row is the coordinates for a vertex.}

\item{Tr}{The triangulation matrix of dimention nT by three, where nT is the
number of triangles in the triangulation. Each row is the indices of vertices in V.}

\item{d}{The degree of piecewise polynomials – default is 5, and usually d
is greater than one. -1 represents piecewise constant.}

\item{r}{The smoothness parameter – default is 1, and 0 ≤ r < d.}

\item{Y}{A vector of response variable.}

\item{X}{A matrix containing the values of model covariates with
spatially varying effects.}

\item{S}{A matrix containing the locations of new points.}

\item{family}{This is a family object specifying the distribution and link
to use.}
}
\description{
mfit.gsvcm(): This function fits a generalized SVCM (GSVCM) on a subregion.
}
