% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gplsvcm.R
\name{plots.gplsvcm}
\alias{plots.gplsvcm}
\title{gplSVCM: generalized partially linear model with spatially varying coefficients}
\usage{
plots.gplsvcm(mfit, S.grid = NULL)
}
\arguments{
\item{mfit}{Fitted ``GPL-SVCM" object}

\item{S.grid}{A matrix containing the locations of points to be evaluated.}
}
\value{
\item{p.all}{A list of rasters plots of estimated coefficient functions.}
}
\description{
This function generates raster plots for the estimated coefficient functions.
}
